Use size_t index to iterate up to std::vector::size() (#1251)
Fixes a different signedness compare warning.

+++ tools/util/include/cutlass/util/command_line.h
@@ -185,7 +185,7 @@ struct CommandLine {
       vals.clear();
 
       // Recover from multi-value string
-      for (int i = 0; i < keys.size(); ++i) {
+      for (size_t i = 0; i < keys.size(); ++i) {
         if (keys[i] == string(arg_name)) {
           string val_string(values[i]);
           separate_string(val_string, vals, sep);
